requirements.txt
Django>=4.2.0
Pillow>=10.0.0
python-dotenv>=1.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-ckeditor>=6.7.0
django-cleanup>=8.0.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.2,<6.0
django-cors-headers>=4.0,<5.0 